Refactoring of directory structure --> Moved all scripts into script folder
