+++ scripts/Curve.py
@@ -2,7 +2,7 @@
 import numpy as np
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
-from HistogramData import HistogramData
+from scripts.HistogramData import HistogramData
 # endregion
